+++ database/migrations/2023_03_13_102522_create_invoice_items_table.php
@@ -13,18 +13,32 @@ return new class extends Migration
      */
     public function up()
     {
+        // Schema::create('invoice_items', function (Blueprint $table) {
+        //     $table->id();
+        //     $table->foreignId('transection_id')->constrained();
+        //     $table->string('item');
+        //     $table->string('size');
+        //     $table->float('width');
+        //     $table->float('height');
+        //     $table->float('square_ft');
+        //     $table->double('qty');
+        //     $table->float('total_square_ft');
+        //     $table->float('rate');
+        //     $table->float('price');
+        //     $table->text('entry_date');
+        // });
         Schema::create('invoice_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('transection_id')->constrained();
             $table->string('item');
             $table->string('size');
-            $table->float('width');
-            $table->float('height');
-            $table->float('square_ft');
-            $table->integer('qty');
-            $table->float('total_square_ft');
-            $table->float('rate');
-            $table->float('price');
+            $table->double('width');
+            $table->double('height');
+            $table->double('square_ft');
+            $table->double('qty');
+            $table->double('total_square_ft');
+            $table->double('rate');
+            $table->double('price');
             $table->text('entry_date');
         });
     }
